chore(user): Adjust service and controller for global pageable
